Restructure the repo. Split generic and specific benchmark code. Add kvs helper
