+++ pyproject.toml
@@ -1,3 +1,7 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
 [project]
 name = "actor_benchmarks"
 version = "0.0.1"
@@ -19,18 +23,16 @@ dependencies = [
 dev = [
     "mypy~=1.15.0",
     "pre-commit~=4.2.0",
-    "pytest-asyncio~=0.26.0",
-    "pytest-cov~=6.1.0",
-    "pytest-only~=2.1.0",
-    "pytest-xdist~=3.6.0",
-    "pytest~=8.3.0",
     "ruff~=0.11.0",
 ]
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/actor_benchmarks"]
 
 [tool.mypy]
 python_version = "3.13"
 plugins = ["pydantic.mypy"]
-files = ["crawler_actors_configurations"]
+files = ["src", "crawler_actors"]
 check_untyped_defs = true
 disallow_incomplete_defs = true
 disallow_untyped_calls = true
@@ -45,4 +47,4 @@ warn_unused_ignores = true
 [tool.basedpyright]
 pythonVersion = "3.13"
 typeCheckingMode = "standard"
-include = ["crawler_actors_configurations"]
+include = ["src", "crawler_actors"]
